requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
boto3==1.28.32
botocore==1.31.32
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.2.0
click==8.1.7
dataclasses-json==0.5.14
decorator==4.4.2
einops==0.6.1
encodec==0.1.1
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
funcy==2.0
googletrans==4.0.0rc1
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.16.4
hyperframe==5.2.0
idna==2.10
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.271
langdetect==1.0.9
langsmith==0.0.26
llvmlite==0.40.1
MarkupSafe==2.1.3
marshmallow==3.20.1
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numba==0.57.1
numexpr==2.8.5
numpy==1.24.4
openai==0.27.9
openai-whisper @ git+https://github.com/openai/whisper.git@e8622f9afc4eba139bf796c210f5c01081000472
packaging==23.1
Pillow==10.0.0
proglog==0.1.10
pydantic==2.2.1
pydantic_core==2.6.1
pydub==0.25.1
python-dateutil==2.8.2
pytube==15.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rfc3986==1.5.0
s3transfer==0.6.2
safetensors==0.3.3
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
suno-bark @ git+https://github.com/suno-ai/bark.git@56b0ba13f7c281cbffa07ea9abf7b30273a60b6a
sympy==1.12
tenacity==8.2.3
tiktoken==0.3.3
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
tqdm==4.66.1
transformers @ git+https://github.com/huggingface/transformers.git@4d40109c3a93c9b8bbca204cb046ed510f1c72e8
typing-inspect==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
yarl==1.9.2
zipp==3.16.2
